requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
beautifulsoup4==4.12.2
requests==2.31.0
django-filter==23.3
drf-yasg==1.21.7
Pillow==10.1.0
gunicorn==21.2.0
setuptools>=65.0.0